refactor(capabilities): make gemini own its kinds and split mod.rs
Move the gemini mail vocabulary out of `aether-kinds` into the cap that
owns it (ADR-0121) and decompose the 1293-line `gemini/mod.rs` along its
seams.

The four `Kind`-deriving request/result types plus their supporting
`Schema`-only enums/structs (AspectRatio, ImageSize, ThinkingLevel,
GroundingMetadata, GeminiError) move verbatim from
`aether-kinds/src/lib.rs` into `gemini/kinds.rs`, names and field schemas
byte-identical so the `fnv1a_64` kind ids are unchanged. `Usage` stays in
`aether-kinds` (shared with the still-central `aether.anthropic` kinds)
and is imported back. Because `aether-capabilities` links into every
chassis binary, `describe_kinds` / `descriptors::all()` keep surfacing
the four kinds; the descriptor-inventory coverage moves with them out of
`aether-kinds` (which can no longer name the types) into a `gemini/kinds.rs`
test block.

`gemini/mod.rs` splits into `config.rs` (the derive-`Config` struct),
`adapter.rs` (the disabled/ureq backends, request encoding, and the
provider string mappers), and `capability.rs` (the `GeminiCapability`
actor and its handlers), each carrying the tests that exercise it; `mod.rs`
is left with the crate doc, module decls, consts, and re-exports.
