+++ crates/aether-capabilities/src/gemini/mod.rs
@@ -17,23 +17,25 @@
 //! A key-absent boot wires [`DisabledGeminiAdapter`] so the mailbox
 //! still loads and replies `Err { Unauthorized }`. CI smokes wire the
 //! `StubGeminiAdapter` from issue 1013.
+//!
+//! Per ADR-0121 the cap owns its mail kinds — the request / result types
+//! and their supporting enums live in `kinds`, decomposed alongside the
+//! `config`, `adapter`, and `capability` seams.
 
+mod adapter;
+mod capability;
+mod config;
 mod error;
+mod kinds;
 mod lyria;
 mod nanobanana;
 
-use std::time::Duration;
-
-use serde_json::Value;
-
-use crate::contentgen::adapter::{
-    AdapterUsage, GeminiAdapter, GeminiArtifact, GeminiImageRequest, GeminiMusicRequest,
-    GeminiResponse,
-};
-use crate::contentgen::shared;
-
 use crate::config_env::DEFAULT_PROVIDER_MAX_IN_FLIGHT;
+
+pub use adapter::{DisabledGeminiAdapter, UreqGeminiAdapter};
+pub use capability::GeminiCapability;
 pub use config::{GeminiConfig, GeminiConfigLayer, GeminiOverlay};
+pub use kinds::*;
 
 /// Default per-cap concurrency bound when `AETHER_GEMINI_MAX_IN_FLIGHT`
 /// is unset. Conservative — image / music generation is multi-second
@@ -43,1251 +45,3 @@ pub const DEFAULT_MAX_IN_FLIGHT: usize = DEFAULT_PROVIDER_MAX_IN_FLIGHT;
 /// Default per-request timeout when `AETHER_GEMINI_TIMEOUT_MS` is unset.
 /// Media generation can run a couple minutes.
 pub const DEFAULT_TIMEOUT_MILLIS: u32 = 180_000;
-
-/// Adapter returned when `GEMINI_API_KEY` is unset (or
-/// `AETHER_GEMINI_DISABLE=1`). Every request replies
-/// `Err { Unauthorized }` so a key-absent boot still loads rather than
-/// warn-dropping.
-pub struct DisabledGeminiAdapter;
-
-impl GeminiAdapter for DisabledGeminiAdapter {
-    fn nanobanana_generate(&self, _req: GeminiImageRequest) -> Result<GeminiResponse, String> {
-        Err(error::UNAUTHORIZED_SENTINEL.to_string())
-    }
-
-    fn lyria_generate(&self, _req: GeminiMusicRequest) -> Result<GeminiResponse, String> {
-        Err(error::UNAUTHORIZED_SENTINEL.to_string())
-    }
-}
-
-mod config {
-    use super::{DEFAULT_MAX_IN_FLIGHT, DEFAULT_TIMEOUT_MILLIS};
-    // confique consumes these through `#[config(parse_env = …)]`; IntelliJ-Rust
-    // doesn't trace macro-attr path args (Qodana FP), but rustc + clippy do.
-    #[allow(unused_imports)]
-    use crate::config_env::{parse_flag, parse_millis_strict, parse_provider_max_in_flight_strict};
-    use std::time::Duration;
-
-    /// Resolved configuration for the `aether.gemini` cap. Chassis mains
-    /// read env (`GEMINI_API_KEY`, `AETHER_GEMINI_DISABLE`,
-    /// `AETHER_GEMINI_MAX_IN_FLIGHT`, `AETHER_GEMINI_TIMEOUT_MS`); the
-    /// staging root reads `AETHER_GEN_DIR` at stage time (shared with
-    /// issue 1013 / 1014). Tests build it directly.
-    ///
-    /// ADR-0090 unit g (iamacoffeepot/aether#1264): the
-    /// `#[derive(aether_substrate::Config)]` emits the env-shaped
-    /// `GeminiConfigLayer`, the clap-shaped `GeminiOverlay`, the
-    /// `FromArgvThenEnv` impl, and the inherent `from_env` shims under
-    /// `feature = "native"`. The wasm-marker build carries only the
-    /// domain struct.
-    #[derive(Clone, Debug)]
-    #[cfg_attr(feature = "native", derive(aether_substrate::Config))]
-    #[cfg_attr(
-        feature = "native",
-        config(env_prefix = "AETHER_GEMINI", cli_prefix = "gemini")
-    )]
-    pub struct GeminiConfig {
-        /// The Google API key. `None` (or `disabled`) wires the
-        /// `DisabledGeminiAdapter`. `env` override pins the unprefixed
-        /// `GEMINI_API_KEY` key.
-        #[cfg_attr(feature = "native", config(env = "GEMINI_API_KEY"))]
-        pub api_key: Option<String>,
-        /// `AETHER_GEMINI_DISABLE=1` forces the disabled adapter.
-        /// `env` + `cli_long` overrides pin the historical wire shape
-        /// (no `D` suffix on `DISABLE`).
-        #[cfg_attr(
-            feature = "native",
-            config(
-                env = "AETHER_GEMINI_DISABLE",
-                cli_long = "gemini-disable",
-                default = false,
-                parse = parse_flag
-            )
-        )]
-        pub disabled: bool,
-        /// Per-cap concurrency bound (doubles as rate-limit throttling).
-        #[cfg_attr(
-            feature = "native",
-            config(default = 2, parse = parse_provider_max_in_flight_strict)
-        )]
-        pub max_in_flight: usize,
-        /// Per-request timeout for the media APIs. The derive's
-        /// `ms_duration` hint + `layer_field = "timeout_ms"` pin the
-        /// Layer / env / CLI shape to the pre-derive name
-        /// (`AETHER_GEMINI_TIMEOUT_MS`, `--gemini-timeout-ms`).
-        #[cfg_attr(
-            feature = "native",
-            config(
-                default = 180_000,
-                parse = parse_millis_strict::<DEFAULT_TIMEOUT_MILLIS>,
-                ms_duration,
-                layer_field = "timeout_ms"
-            )
-        )]
-        pub timeout: Duration,
-    }
-
-    impl Default for GeminiConfig {
-        fn default() -> Self {
-            Self {
-                api_key: None,
-                disabled: false,
-                max_in_flight: DEFAULT_MAX_IN_FLIGHT,
-                timeout: Duration::from_millis(u64::from(DEFAULT_TIMEOUT_MILLIS)),
-            }
-        }
-    }
-
-    // confique's `parse_env` contract is `fn(&str) -> Result<T, impl Error>`,
-    // so these total helpers carry a `Result` they never fill with `Err`.
-    // The strict (erroring) variants land with the ADR-0090 §4 validation
-    // pass; hence the per-fn `unnecessary_wraps` allow.
-
-    #[cfg(all(test, feature = "native"))]
-    mod tests {
-        use super::{
-            DEFAULT_MAX_IN_FLIGHT, DEFAULT_TIMEOUT_MILLIS, GeminiConfig, GeminiConfigLayer,
-        };
-        use confique::Config as _;
-        use std::time::Duration;
-
-        // ADR-0090: the confique migration is byte-identical to the prior
-        // hand-rolled reader. These exercise the resolution logic without
-        // touching process env (issue 464).
-
-        #[test]
-        fn gemini_from_env_defaults_match() {
-            // No `.env()` source: loads literal defaults only, so this is
-            // env-free and guards the layer's defaults against the named
-            // consts + `GeminiConfig::default()`.
-            let layer = GeminiConfigLayer::builder().load().expect("defaults load");
-            let default = GeminiConfig::default();
-            assert_eq!(layer.api_key, None);
-            assert!(!layer.disabled);
-            assert_eq!(layer.max_in_flight, DEFAULT_MAX_IN_FLIGHT);
-            assert_eq!(layer.timeout_ms, DEFAULT_TIMEOUT_MILLIS);
-            assert_eq!(
-                Duration::from_millis(u64::from(layer.timeout_ms)),
-                default.timeout
-            );
-        }
-    }
-}
-
-/// `ureq`-backed Gemini media backend. Holds the shared agent, the API
-/// key, and the per-request timeout. The blocking calls run on the
-/// spawn-and-die ephemeral thread.
-pub struct UreqGeminiAdapter {
-    agent: ureq::Agent,
-    api_key: String,
-    timeout: Duration,
-}
-
-impl UreqGeminiAdapter {
-    /// Build the adapter with a resolved key + timeout.
-    #[must_use]
-    pub fn new(api_key: String, timeout: Duration) -> Self {
-        Self {
-            agent: shared::agent(),
-            api_key,
-            timeout,
-        }
-    }
-}
-
-/// Generative Language API host.
-const GENLANG_HOST: &str = "https://generativelanguage.googleapis.com";
-
-impl UreqGeminiAdapter {
-    /// POST a JSON body to a Generative Language API endpoint and return
-    /// the response text on a 2xx. Both media backends share this shape
-    /// (build request → run → status-check), so it lives in one place.
-    /// `endpoint` is the `:method` suffix (`generateContent` / `predict`).
-    fn post_json(&self, model: &str, endpoint: &str, body: &Value) -> Result<String, String> {
-        use ureq::http::Request;
-        let body_bytes = serde_json::to_vec(body).map_err(|e| format!("encode request: {e}"))?;
-        let url = format!("{GENLANG_HOST}/v1beta/models/{model}:{endpoint}");
-        let http_req = Request::builder()
-            .method("POST")
-            .uri(&url)
-            .header("x-goog-api-key", &self.api_key)
-            .header("content-type", "application/json")
-            .body(body_bytes)
-            .map_err(|e| format!("build request: {e}"))?;
-        let (status, retry_after_millis, text) =
-            shared::run_request(&self.agent, http_req, self.timeout)?;
-        if !(200..300).contains(&status) {
-            return Err(format!(
-                "status={status} retry_after_ms={retry_after_millis:?} body={text}"
-            ));
-        }
-        Ok(text)
-    }
-}
-
-/// Build the `generateContent` request body for a Nano Banana image
-/// request: the prompt plus any reference images as inline-data parts,
-/// and the per-request knobs in `generationConfig`. Each optional field
-/// is emitted only when its source `Option` is `Some(..)` so an unset
-/// knob leaves no key in the body. Factored out so a unit test can lock
-/// the JSON shape without an HTTP call.
-fn build_nanobanana_body(req: &GeminiImageRequest) -> Value {
-    use serde_json::{Map, json};
-
-    let mut parts = vec![json!({ "text": req.prompt })];
-    for img in &req.reference_images {
-        parts.push(json!({
-            "inlineData": {
-                "mimeType": "image/png",
-                "data": base64_encode(img),
-            }
-        }));
-    }
-
-    // `imageConfig` always carries the aspect ratio; `imageSize` rides
-    // alongside it under the same object when set (issue 1167 — do not
-    // switch to `responseFormat.image`).
-    let mut image_config = Map::new();
-    image_config.insert("aspectRatio".to_string(), json!(req.aspect_ratio));
-    if let Some(size) = &req.image_size {
-        image_config.insert("imageSize".to_string(), json!(size));
-    }
-
-    let mut generation_config = Map::new();
-    generation_config.insert("imageConfig".to_string(), Value::Object(image_config));
-
-    // `thinkingConfig` only appears when at least one of its fields is
-    // set; each field is emitted independently.
-    let mut thinking_config = Map::new();
-    if let Some(level) = &req.thinking_level {
-        thinking_config.insert("thinkingLevel".to_string(), json!(level));
-    }
-    if let Some(include) = req.include_thoughts {
-        thinking_config.insert("includeThoughts".to_string(), json!(include));
-    }
-    if !thinking_config.is_empty() {
-        generation_config.insert("thinkingConfig".to_string(), Value::Object(thinking_config));
-    }
-
-    let mut body = Map::new();
-    body.insert(
-        "contents".to_string(),
-        json!([{ "role": "user", "parts": parts }]),
-    );
-    body.insert(
-        "generationConfig".to_string(),
-        Value::Object(generation_config),
-    );
-    if req.use_grounding {
-        body.insert("tools".to_string(), json!([{ "google_search": {} }]));
-    }
-
-    Value::Object(body)
-}
-
-impl GeminiAdapter for UreqGeminiAdapter {
-    fn nanobanana_generate(&self, req: GeminiImageRequest) -> Result<GeminiResponse, String> {
-        let body = build_nanobanana_body(&req);
-        let text = self.post_json(&req.model, "generateContent", &body)?;
-
-        let parsed = nanobanana::parse_image_response(&text)?;
-        Ok(GeminiResponse {
-            artifacts: vec![GeminiArtifact {
-                bytes: parsed.bytes,
-                ext: "png".to_string(),
-            }],
-            model_used: req.model,
-            usage: AdapterUsage::default(),
-            thought_signature: parsed.thought_signature,
-            grounding: parsed.grounding,
-        })
-    }
-
-    fn lyria_generate(&self, req: GeminiMusicRequest) -> Result<GeminiResponse, String> {
-        use serde_json::json;
-
-        let body = json!({
-            "instances": [{ "prompt": req.prompt }],
-            "parameters": { "sampleCount": req.sample_count.max(1) },
-        });
-        let text = self.post_json(&req.model, "predict", &body)?;
-
-        let clips = lyria::parse_clip_response(&text)?;
-        let artifacts = clips
-            .into_iter()
-            .map(|bytes| GeminiArtifact {
-                bytes,
-                ext: "wav".to_string(),
-            })
-            .collect();
-        Ok(GeminiResponse {
-            artifacts,
-            model_used: req.model,
-            usage: AdapterUsage::default(),
-            thought_signature: None,
-            grounding: None,
-        })
-    }
-}
-
-/// Minimal standard-alphabet base64 encoder for reference-image bytes
-/// on the request side (no padding omitted). Avoids a base64 crate.
-fn base64_encode(bytes: &[u8]) -> String {
-    const ALPHABET: &[u8; 64] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-    let mut out = String::with_capacity(bytes.len().div_ceil(3) * 4);
-    for chunk in bytes.chunks(3) {
-        let b0 = u32::from(chunk[0]);
-        let b1 = chunk.get(1).copied().map_or(0, u32::from);
-        let b2 = chunk.get(2).copied().map_or(0, u32::from);
-        let n = (b0 << 16) | (b1 << 8) | b2;
-        out.push(ALPHABET[((n >> 18) & 0x3F) as usize] as char);
-        out.push(ALPHABET[((n >> 12) & 0x3F) as usize] as char);
-        out.push(if chunk.len() > 1 {
-            ALPHABET[((n >> 6) & 0x3F) as usize] as char
-        } else {
-            '='
-        });
-        out.push(if chunk.len() > 2 {
-            ALPHABET[(n & 0x3F) as usize] as char
-        } else {
-            '='
-        });
-    }
-    out
-}
-
-/// Map the wire `AspectRatio` to the provider's `W:H` string.
-fn aspect_ratio_str(ar: aether_kinds::AspectRatio) -> &'static str {
-    use aether_kinds::AspectRatio as A;
-    match ar {
-        A::ASPECT_RATIO_1_1 => "1:1",
-        A::ASPECT_RATIO_2_3 => "2:3",
-        A::ASPECT_RATIO_3_2 => "3:2",
-        A::ASPECT_RATIO_3_4 => "3:4",
-        A::ASPECT_RATIO_4_3 => "4:3",
-        A::ASPECT_RATIO_4_5 => "4:5",
-        A::ASPECT_RATIO_5_4 => "5:4",
-        A::ASPECT_RATIO_9_16 => "9:16",
-        A::ASPECT_RATIO_16_9 => "16:9",
-        A::ASPECT_RATIO_21_9 => "21:9",
-        A::ASPECT_RATIO_1_4 => "1:4",
-        A::ASPECT_RATIO_1_8 => "1:8",
-        A::ASPECT_RATIO_4_1 => "4:1",
-        A::ASPECT_RATIO_8_1 => "8:1",
-    }
-}
-
-/// Map the wire `ImageSize` to the provider's `imageConfig.imageSize`
-/// string. Uppercase `K`; `"512"` has no `K`.
-fn image_size_str(size: aether_kinds::ImageSize) -> &'static str {
-    use aether_kinds::ImageSize as S;
-    match size {
-        S::S512 => "512",
-        S::K1 => "1K",
-        S::K2 => "2K",
-        S::K4 => "4K",
-    }
-}
-
-/// Map the wire `ThinkingLevel` to the provider's
-/// `thinkingConfig.thinkingLevel` string.
-fn thinking_level_str(level: aether_kinds::ThinkingLevel) -> &'static str {
-    use aether_kinds::ThinkingLevel as T;
-    match level {
-        T::Minimal => "minimal",
-        T::High => "high",
-    }
-}
-
-// Handler-signature kinds must be importable at file root because
-// `#[bridge]` emits `impl HandlesKind<K> for X {}` markers as siblings
-// of the mod.
-use aether_kinds::{
-    LyriaGenerate, LyriaGenerateResult, NanobananaGenerate, NanobananaGenerateResult,
-};
-
-/// Convert an adapter error string into the typed `GeminiError`.
-fn map_adapter_error(raw: &str) -> aether_kinds::GeminiError {
-    error::adapter_error_to_typed(raw)
-}
-
-#[aether_actor::bridge(singleton)]
-mod native {
-    use std::sync::Arc;
-
-    use super::{
-        DisabledGeminiAdapter, GeminiAdapter, GeminiConfig, GeminiImageRequest, GeminiMusicRequest,
-        LyriaGenerate, LyriaGenerateResult, NanobananaGenerate, NanobananaGenerateResult,
-        UreqGeminiAdapter, aspect_ratio_str, image_size_str, lyria, map_adapter_error, nanobanana,
-        thinking_level_str,
-    };
-    use crate::contentgen::adapter::{AdapterUsage, GeminiResponse};
-    use crate::contentgen::staging::{gen_root, stage_gen_output};
-    use crate::contentgen::task_queue::TaskQueue;
-    use crate::fs::{FileAdapter, LocalFileAdapter};
-    use aether_actor::{Manual, OutboundReply, actor};
-    use aether_kinds::{GeminiError, GroundingMetadata, Usage};
-    use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx, TaskDone};
-    use aether_substrate::chassis::error::BootError;
-
-    /// `aether.gemini` mailbox cap. Owns the resolved adapter + the
-    /// cap-level rate-limit queue over the ADR-0093 dispatch primitive.
-    /// Single-threaded post-ADR-0038, so the queue state lives in plain
-    /// fields.
-    pub struct GeminiCapability {
-        adapter: Arc<dyn GeminiAdapter>,
-        tasks: TaskQueue,
-    }
-
-    #[cfg(test)]
-    impl GeminiCapability {
-        pub(crate) fn from_parts(adapter: Arc<dyn GeminiAdapter>, max_in_flight: usize) -> Self {
-            Self {
-                adapter,
-                tasks: TaskQueue::new(max_in_flight),
-            }
-        }
-
-        pub(crate) fn test_in_flight(&self) -> usize {
-            self.tasks.in_flight()
-        }
-    }
-
-    fn build_adapter(config: &GeminiConfig) -> Arc<dyn GeminiAdapter> {
-        if config.disabled {
-            tracing::info!(
-                target: "aether_capabilities::gemini",
-                "gemini adapter disabled — every request replies Unauthorized",
-            );
-            return Arc::new(DisabledGeminiAdapter);
-        }
-        config.api_key.as_ref().map_or_else(
-            || {
-                tracing::info!(
-                    target: "aether_capabilities::gemini",
-                    "GEMINI_API_KEY unset — every request replies Unauthorized",
-                );
-                Arc::new(DisabledGeminiAdapter) as Arc<dyn GeminiAdapter>
-            },
-            |key| {
-                tracing::info!(
-                    target: "aether_capabilities::gemini",
-                    "gemini adapter configured (nanobanana + lyria)",
-                );
-                Arc::new(UreqGeminiAdapter::new(key.clone(), config.timeout))
-                    as Arc<dyn GeminiAdapter>
-            },
-        )
-    }
-
-    /// Read reference-image bytes from the supplied save-namespace
-    /// paths (tool JSON takes paths, the wire stays bytes —
-    /// `feedback_no_bytes_in_llm_json`). A read failure aborts the
-    /// request with an `AdapterError`.
-    fn read_reference_images(paths: &[String]) -> Result<Vec<Vec<u8>>, GeminiError> {
-        if paths.is_empty() {
-            return Ok(Vec::new());
-        }
-        let root = gen_root();
-        let adapter = LocalFileAdapter::new(root, true)
-            .map_err(|e| GeminiError::AdapterError(e.to_string()))?;
-        let mut out = Vec::with_capacity(paths.len());
-        for path in paths {
-            let bytes = adapter
-                .read(path)
-                .map_err(|e| GeminiError::AdapterError(format!("reference {path}: {e:?}")))?;
-            out.push(bytes);
-        }
-        Ok(out)
-    }
-
-    fn to_usage(u: AdapterUsage) -> Usage {
-        Usage {
-            input_tokens: u.input_tokens,
-            output_tokens: u.output_tokens,
-            wall_clock_ms: u.wall_clock_ms,
-            cost_micros: u.cost_micros,
-        }
-    }
-
-    fn nanobanana_reply(
-        request_id: u64,
-        include_sig: bool,
-        result: Result<GeminiResponse, String>,
-    ) -> NanobananaGenerateResult {
-        match result {
-            Ok(resp) => {
-                let model_used = resp.model_used;
-                let usage = to_usage(resp.usage);
-                // Opt-in / default-off: clear the signature unless the
-                // caller asked to retain it for a multi-turn continuation
-                // (a signature can run to multiple MB and dominate the
-                // reply). Parse stays unconditional; the gate is here.
-                let thought_signature = if include_sig {
-                    resp.thought_signature
-                } else {
-                    None
-                };
-                let grounding =
-                    resp.grounding
-                        .map(|(search_queries, source_urls)| GroundingMetadata {
-                            search_queries,
-                            source_urls,
-                        });
-                let Some(artifact) = resp.artifacts.into_iter().next() else {
-                    return NanobananaGenerateResult::Err {
-                        request_id,
-                        error: GeminiError::AdapterError("adapter returned no image".to_string()),
-                    };
-                };
-                match stage_gen_output(&artifact.bytes, &artifact.ext) {
-                    Ok(output_path) => NanobananaGenerateResult::Ok {
-                        request_id,
-                        output_path,
-                        model_used,
-                        usage,
-                        thought_signature,
-                        grounding,
-                    },
-                    Err(e) => NanobananaGenerateResult::Err {
-                        request_id,
-                        error: GeminiError::AdapterError(format!("stage image: {e:?}")),
-                    },
-                }
-            }
-            Err(raw) => NanobananaGenerateResult::Err {
-                request_id,
-                error: map_adapter_error(&raw),
-            },
-        }
-    }
-
-    fn lyria_reply(request_id: u64, result: Result<GeminiResponse, String>) -> LyriaGenerateResult {
-        match result {
-            Ok(resp) => {
-                let mut output_paths = Vec::with_capacity(resp.artifacts.len());
-                for artifact in &resp.artifacts {
-                    match stage_gen_output(&artifact.bytes, &artifact.ext) {
-                        Ok(path) => output_paths.push(path),
-                        Err(e) => {
-                            return LyriaGenerateResult::Err {
-                                request_id,
-                                error: GeminiError::AdapterError(format!("stage clip: {e:?}")),
-                            };
-                        }
-                    }
-                }
-                LyriaGenerateResult::Ok {
-                    request_id,
-                    output_paths,
-                    model_used: resp.model_used,
-                    usage: to_usage(resp.usage),
-                }
-            }
-            Err(raw) => LyriaGenerateResult::Err {
-                request_id,
-                error: map_adapter_error(&raw),
-            },
-        }
-    }
-
-    #[actor]
-    impl NativeActor for GeminiCapability {
-        type Config = GeminiConfig;
-
-        /// ADR-0050 + ADR-0074 Phase 5 chassis-owned mailbox.
-        const NAMESPACE: &'static str = "aether.gemini";
-
-        fn init(config: GeminiConfig, _ctx: &mut NativeInitCtx<'_>) -> Result<Self, BootError> {
-            Ok(Self {
-                adapter: build_adapter(&config),
-                tasks: TaskQueue::new(config.max_in_flight),
-            })
-        }
-
-        /// Generate an image via Nano Banana off the dispatcher thread.
-        ///
-        /// # Agent
-        /// Reply: `NanobananaGenerateResult` carrying a staged
-        /// `save://gen/<uuid>.png` path. Validates the model and the
-        /// per-model `aspect_ratio` / `image_size` / reference-count
-        /// rules synchronously (the matching `…NotSupportedByModel` /
-        /// `UnknownModel` error on a miss) before any dispatch.
-        #[handler::manual]
-        fn on_nanobanana_generate(
-            &mut self,
-            ctx: &mut NativeCtx<'_, Manual>,
-            mail: NanobananaGenerate,
-        ) {
-            let request_id = mail.request_id;
-            // Opt-in / default-off; cross-model, so never validated.
-            let include_sig = mail.include_thought_signature.unwrap_or(false);
-            let Some(shape) = nanobanana::lookup_model(&mail.model) else {
-                OutboundReply::reply(
-                    ctx,
-                    &NanobananaGenerateResult::Err {
-                        request_id,
-                        error: GeminiError::UnknownModel {
-                            model: mail.model,
-                            supported: nanobanana::supported_model_ids(),
-                        },
-                    },
-                );
-                return;
-            };
-            let inputs = nanobanana::ValidationInputs {
-                aspect_ratio: mail.aspect_ratio,
-                image_size: mail.image_size,
-                thinking_level_set: mail.thinking_level.is_some(),
-                include_thoughts_set: mail.include_thoughts.is_some(),
-                use_grounding_set: mail.use_grounding.is_some(),
-                object_ref_count: mail.object_reference_paths.len(),
-                character_ref_count: mail.character_reference_paths.len(),
-            };
-            if let Err(error) = nanobanana::validate(shape, &inputs) {
-                OutboundReply::reply(ctx, &NanobananaGenerateResult::Err { request_id, error });
-                return;
-            }
-
-            // Read reference bytes on the dispatcher thread (small,
-            // local) before handing the network call off-thread.
-            let mut ref_paths = mail.object_reference_paths;
-            ref_paths.extend(mail.character_reference_paths);
-            let reference_images = match read_reference_images(&ref_paths) {
-                Ok(b) => b,
-                Err(error) => {
-                    OutboundReply::reply(ctx, &NanobananaGenerateResult::Err { request_id, error });
-                    return;
-                }
-            };
-
-            let req = GeminiImageRequest {
-                model: mail.model,
-                prompt: mail.prompt,
-                aspect_ratio: aspect_ratio_str(mail.aspect_ratio).to_string(),
-                image_size: mail.image_size.map(|s| image_size_str(s).to_string()),
-                thinking_level: mail
-                    .thinking_level
-                    .map(|l| thinking_level_str(l).to_string()),
-                include_thoughts: mail.include_thoughts,
-                use_grounding: mail.use_grounding.unwrap_or(false),
-                reference_images,
-            };
-            let adapter = Arc::clone(&self.adapter);
-            self.tasks.submit(ctx, move || {
-                let result = adapter.nanobanana_generate(req);
-                // Staging runs here on the worker (blocking disk I/O), so
-                // a megabyte PNG never rides the mail wire — the reply
-                // carries the staged path.
-                nanobanana_reply(request_id, include_sig, result)
-            });
-        }
-
-        /// Generate music via Lyria off the dispatcher thread.
-        ///
-        /// # Agent
-        /// Reply: `LyriaGenerateResult` carrying one staged
-        /// `save://gen/<uuid>.wav` path per clip. Rejects an unknown
-        /// model and a both-set `seed` + `sample_count` synchronously
-        /// before any dispatch.
-        #[handler::manual]
-        fn on_lyria_generate(&mut self, ctx: &mut NativeCtx<'_, Manual>, mail: LyriaGenerate) {
-            let request_id = mail.request_id;
-            if !lyria::is_supported(&mail.model) {
-                OutboundReply::reply(
-                    ctx,
-                    &LyriaGenerateResult::Err {
-                        request_id,
-                        error: GeminiError::UnknownModel {
-                            model: mail.model,
-                            supported: lyria::supported_model_ids(),
-                        },
-                    },
-                );
-                return;
-            }
-            if let Err(error) = lyria::validate(
-                &mail.model,
-                mail.seed.is_some(),
-                mail.sample_count.is_some(),
-            ) {
-                OutboundReply::reply(ctx, &LyriaGenerateResult::Err { request_id, error });
-                return;
-            }
-
-            let req = GeminiMusicRequest {
-                model: mail.model,
-                prompt: mail.prompt,
-                sample_count: mail.sample_count.unwrap_or(1),
-            };
-            let adapter = Arc::clone(&self.adapter);
-            self.tasks.submit(ctx, move || {
-                let result = adapter.lyria_generate(req);
-                // Staging (one path per clip) runs here on the worker.
-                lyria_reply(request_id, result)
-            });
-        }
-
-        /// ADR-0093 completion for a finished Nano Banana call: re-reply
-        /// the worker's staged result to the original caller (drops the
-        /// hold), then free the in-flight slot (draining the next pending
-        /// request).
-        #[handler(task)]
-        fn on_nanobanana_done(
-            &mut self,
-            ctx: &mut NativeCtx<'_>,
-            done: TaskDone<NanobananaGenerateResult>,
-        ) {
-            done.resolve(ctx);
-            self.tasks.on_complete(ctx);
-        }
-
-        /// ADR-0093 completion for a finished Lyria call.
-        #[handler(task)]
-        fn on_lyria_done(&mut self, ctx: &mut NativeCtx<'_>, done: TaskDone<LyriaGenerateResult>) {
-            done.resolve(ctx);
-            self.tasks.on_complete(ctx);
-        }
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::super::{DisabledGeminiAdapter, GeminiConfig};
-        use super::GeminiCapability;
-        use crate::contentgen::adapter::STUB_PNG;
-        use crate::contentgen::adapter::StubGeminiAdapter;
-        use crate::contentgen::adapter::{AdapterUsage, GeminiArtifact, GeminiResponse};
-        use crate::test_chassis::{
-            TestChassis, decode_session_reply, drive_task_completion, fresh_substrate,
-            test_mailer_and_rx,
-        };
-        use aether_actor::Addressable;
-        use aether_data::{Kind, MailboxId, SessionToken, Source, SourceAddr, Uuid};
-        use aether_kinds::{
-            AspectRatio, GeminiError, ImageSize, LyriaGenerate, LyriaGenerateResult,
-            NanobananaGenerate, NanobananaGenerateResult,
-        };
-        use aether_substrate::actor::native::binding::NativeBinding;
-        use aether_substrate::actor::native::ctx::NativeCtx;
-        use aether_substrate::chassis::builder::Builder;
-        use aether_substrate::mail::outbound::EgressEvent;
-        use serde::de::DeserializeOwned;
-        use std::sync::mpsc::Receiver;
-        use std::sync::{Arc, Mutex, PoisonError};
-        use std::time::{Duration, SystemTime, UNIX_EPOCH};
-        use std::{env, fs, process};
-
-        fn session_sender() -> Source {
-            Source::to(SourceAddr::Session(SessionToken(Uuid::nil())))
-        }
-
-        /// Thin alias over the shared `decode_session_reply`.
-        fn decode_reply<K: Kind + DeserializeOwned>(rx: &Receiver<EgressEvent>) -> K {
-            decode_session_reply(rx)
-        }
-
-        fn nb_request(model: &str, aspect_ratio: AspectRatio) -> NanobananaGenerate {
-            NanobananaGenerate {
-                request_id: 1,
-                model: model.to_string(),
-                prompt: "a cat".to_string(),
-                aspect_ratio,
-                image_size: None,
-                thinking_level: None,
-                include_thoughts: None,
-                object_reference_paths: Vec::new(),
-                character_reference_paths: Vec::new(),
-                use_grounding: None,
-                include_thought_signature: None,
-            }
-        }
-
-        #[test]
-        fn capability_boots_and_registers_mailbox() {
-            let (registry, mailer) = fresh_substrate();
-            let chassis = Builder::<TestChassis>::new(Arc::clone(&registry), Arc::clone(&mailer))
-                .with_actor::<GeminiCapability>(GeminiConfig::default())
-                .build_passive()
-                .expect("gemini capability boots");
-            assert!(
-                registry.lookup(GeminiCapability::NAMESPACE).is_some(),
-                "gemini mailbox registered"
-            );
-            drop(chassis);
-        }
-
-        /// End-to-end through the ADR-0093 dispatch primitive: the stub
-        /// Nano Banana adapter runs on the real worker thread, stages a PNG
-        /// under a scratch `AETHER_GEN_DIR`, and the cap's
-        /// `#[handler(task)]` completion re-replies the `Ok` result —
-        /// carrying a staged `gen/<uuid>.png` path that exists on disk — to
-        /// the original caller.
-        #[test]
-        fn gemini_stub_nanobanana() {
-            let _guard = GEN_DIR_ENV_LOCK
-                .lock()
-                .unwrap_or_else(PoisonError::into_inner);
-            let scratch = env::temp_dir().join(format!(
-                "aether-gemini-nb-{}-{}",
-                process::id(),
-                request_nonce()
-            ));
-            fs::create_dir_all(&scratch).expect("scratch dir creates");
-            // SAFETY: serialized by GEN_DIR_ENV_LOCK against the other
-            // gen-dir tests; the real worker stages here.
-            unsafe {
-                env::set_var("AETHER_GEN_DIR", &scratch);
-            }
-
-            let (mailer, rx) = test_mailer_and_rx();
-            let cap_mailbox = MailboxId(0);
-            let mut cap = GeminiCapability::from_parts(Arc::new(StubGeminiAdapter), 4);
-            let transport = Arc::new(NativeBinding::new_for_test(
-                Arc::clone(&mailer),
-                cap_mailbox,
-            ));
-            let mut ctx = NativeCtx::new_dispatching(
-                &transport,
-                session_sender(),
-                aether_data::MailId::NONE,
-                aether_data::MailId::NONE,
-            );
-            cap.on_nanobanana_generate(
-                &mut ctx,
-                nb_request(
-                    "gemini-3.1-flash-image-preview",
-                    AspectRatio::ASPECT_RATIO_1_1,
-                ),
-            );
-            // The worker runs the stub call + staging and pushes the
-            // completion wake; route it through the cap's task handler.
-            drive_task_completion(&mut cap, &transport, &rx);
-
-            match decode_reply::<NanobananaGenerateResult>(&rx) {
-                NanobananaGenerateResult::Ok { output_path, .. } => {
-                    assert!(
-                        output_path.starts_with("gen/"),
-                        "staged path was {output_path:?}"
-                    );
-                    assert_eq!(output_path.rsplit('.').next(), Some("png"));
-                    let bytes =
-                        fs::read(scratch.join(&output_path)).expect("staged file exists on disk");
-                    assert_eq!(&bytes[..8], &STUB_PNG[..8]);
-                }
-                other @ NanobananaGenerateResult::Err { .. } => {
-                    panic!("expected Ok, got {other:?}")
-                }
-            }
-
-            // SAFETY: same lock-guarded window as the set above.
-            unsafe {
-                env::remove_var("AETHER_GEN_DIR");
-            }
-            let _ = fs::remove_dir_all(&scratch);
-        }
-
-        /// Per-model validation: an unsupported aspect ratio / image
-        /// size / over-count reference combo errors before any dispatch.
-        #[test]
-        fn gemini_nanobanana_per_model_validation() {
-            let (mailer, rx) = test_mailer_and_rx();
-            let cap_mailbox = MailboxId(0);
-            let mut cap = GeminiCapability::from_parts(Arc::new(StubGeminiAdapter), 4);
-            let transport = Arc::new(NativeBinding::new_for_test(
-                Arc::clone(&mailer),
-                cap_mailbox,
-            ));
-            let mut ctx = NativeCtx::new_dispatching(
-                &transport,
-                session_sender(),
-                aether_data::MailId::NONE,
-                aether_data::MailId::NONE,
-            );
-            // NB1 + the NB2-only extreme aspect ratio -> rejected.
-            cap.on_nanobanana_generate(
-                &mut ctx,
-                nb_request("gemini-2.5-flash-image", AspectRatio::ASPECT_RATIO_8_1),
-            );
-            match decode_reply::<NanobananaGenerateResult>(&rx) {
-                NanobananaGenerateResult::Err {
-                    error: GeminiError::AspectRatioNotSupportedByModel { .. },
-                    ..
-                } => {}
-                other => panic!("expected AspectRatioNotSupportedByModel, got {other:?}"),
-            }
-            // No dispatch happened — the synchronous validation error
-            // never spawned work.
-            assert_eq!(cap.test_in_flight(), 0);
-
-            // NB1 + an unsupported image size -> rejected.
-            let mut req = nb_request("gemini-2.5-flash-image", AspectRatio::ASPECT_RATIO_1_1);
-            req.image_size = Some(ImageSize::S512);
-            cap.on_nanobanana_generate(&mut ctx, req);
-            match decode_reply::<NanobananaGenerateResult>(&rx) {
-                NanobananaGenerateResult::Err {
-                    error: GeminiError::ImageSizeNotSupportedByModel { .. },
-                    ..
-                } => {}
-                other => panic!("expected ImageSizeNotSupportedByModel, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn gemini_unknown_model_errors() {
-            let (mailer, rx) = test_mailer_and_rx();
-            let cap_mailbox = MailboxId(0);
-            let mut cap = GeminiCapability::from_parts(Arc::new(StubGeminiAdapter), 4);
-            let transport = Arc::new(NativeBinding::new_for_test(
-                Arc::clone(&mailer),
-                cap_mailbox,
-            ));
-            let mut ctx = NativeCtx::new_dispatching(
-                &transport,
-                session_sender(),
-                aether_data::MailId::NONE,
-                aether_data::MailId::NONE,
-            );
-            cap.on_nanobanana_generate(
-                &mut ctx,
-                nb_request("gemini-bogus", AspectRatio::ASPECT_RATIO_1_1),
-            );
-            match decode_reply::<NanobananaGenerateResult>(&rx) {
-                NanobananaGenerateResult::Err {
-                    error: GeminiError::UnknownModel { model, supported },
-                    ..
-                } => {
-                    assert_eq!(model, "gemini-bogus");
-                    assert!(supported.contains(&"gemini-3.1-flash-image-preview".to_string()));
-                }
-                other => panic!("expected UnknownModel, got {other:?}"),
-            }
-        }
-
-        /// Lyria stub runs on the real worker, stages WAV clips under a
-        /// scratch `AETHER_GEN_DIR`, and the `#[handler(task)]` completion
-        /// re-replies an `Ok` carrying one staged path per clip.
-        #[test]
-        fn gemini_stub_lyria() {
-            let _guard = GEN_DIR_ENV_LOCK
-                .lock()
-                .unwrap_or_else(PoisonError::into_inner);
-            let scratch = env::temp_dir().join(format!(
-                "aether-gemini-lyria-{}-{}",
-                process::id(),
-                request_nonce()
-            ));
-            fs::create_dir_all(&scratch).expect("scratch dir creates");
-            // SAFETY: serialized by GEN_DIR_ENV_LOCK against the other
-            // gen-dir tests; the real worker stages here.
-            unsafe {
-                env::set_var("AETHER_GEN_DIR", &scratch);
-            }
-
-            let (mailer, rx) = test_mailer_and_rx();
-            let cap_mailbox = MailboxId(0);
-            let mut cap = GeminiCapability::from_parts(Arc::new(StubGeminiAdapter), 4);
-            let transport = Arc::new(NativeBinding::new_for_test(
-                Arc::clone(&mailer),
-                cap_mailbox,
-            ));
-            let mut ctx = NativeCtx::new_dispatching(
-                &transport,
-                session_sender(),
-                aether_data::MailId::NONE,
-                aether_data::MailId::NONE,
-            );
-            cap.on_lyria_generate(
-                &mut ctx,
-                LyriaGenerate {
-                    request_id: 2,
-                    model: "lyria-3".to_string(),
-                    prompt: "ambient".to_string(),
-                    negative_prompt: None,
-                    seed: None,
-                    sample_count: Some(2),
-                },
-            );
-            // The worker runs the stub call + per-clip staging and pushes
-            // the completion wake; route it through the cap's task handler.
-            drive_task_completion(&mut cap, &transport, &rx);
-            match decode_reply::<LyriaGenerateResult>(&rx) {
-                LyriaGenerateResult::Ok { output_paths, .. } => {
-                    assert_eq!(output_paths.len(), 2);
-                    assert!(
-                        output_paths
-                            .iter()
-                            .all(|p| p.starts_with("gen/") && p.rsplit('.').next() == Some("wav"))
-                    );
-                }
-                other @ LyriaGenerateResult::Err { .. } => panic!("expected Ok, got {other:?}"),
-            }
-
-            // SAFETY: same lock-guarded window as the set above.
-            unsafe {
-                env::remove_var("AETHER_GEN_DIR");
-            }
-            let _ = fs::remove_dir_all(&scratch);
-        }
-
-        #[test]
-        fn gemini_disabled_replies_unauthorized() {
-            let (mailer, rx) = test_mailer_and_rx();
-            let cap_mailbox = MailboxId(0);
-            let mut cap = GeminiCapability::from_parts(Arc::new(DisabledGeminiAdapter), 4);
-            let transport = Arc::new(NativeBinding::new_for_test(
-                Arc::clone(&mailer),
-                cap_mailbox,
-            ));
-            let mut ctx = NativeCtx::new_dispatching(
-                &transport,
-                session_sender(),
-                aether_data::MailId::NONE,
-                aether_data::MailId::NONE,
-            );
-            cap.on_nanobanana_generate(
-                &mut ctx,
-                nb_request(
-                    "gemini-3.1-flash-image-preview",
-                    AspectRatio::ASPECT_RATIO_1_1,
-                ),
-            );
-            // The disabled adapter returns the Unauthorized sentinel on the
-            // worker; the completion re-replies the mapped error.
-            drive_task_completion(&mut cap, &transport, &rx);
-            match decode_reply::<NanobananaGenerateResult>(&rx) {
-                NanobananaGenerateResult::Err {
-                    error: GeminiError::Unauthorized,
-                    ..
-                } => {}
-                other => panic!("expected Unauthorized, got {other:?}"),
-            }
-        }
-
-        /// Real-API smoke. Hits the live Nano Banana endpoint with a
-        /// tiny request — ignored by default so CI stays zero-cost; run
-        /// with `GEMINI_API_KEY` set.
-        #[test]
-        #[ignore = "needs GEMINI_API_KEY"]
-        fn gemini_nanobanana_smoke() {
-            use super::super::UreqGeminiAdapter;
-            use crate::contentgen::adapter::{GeminiAdapter, GeminiImageRequest};
-            // Test-only: the live-API smoke reads an external credential
-            // (GEMINI_API_KEY), not cap config; gated `#[ignore]`.
-            #[allow(clippy::disallowed_methods)]
-            let key = env::var("GEMINI_API_KEY").expect("GEMINI_API_KEY set for smoke");
-            let adapter = UreqGeminiAdapter::new(key, Duration::from_mins(1));
-            let resp = adapter
-                .nanobanana_generate(GeminiImageRequest {
-                    model: "gemini-3.1-flash-image-preview".to_string(),
-                    prompt: "a single red dot on white".to_string(),
-                    aspect_ratio: "1:1".to_string(),
-                    reference_images: Vec::new(),
-                    ..Default::default()
-                })
-                .expect("live nanobanana request succeeds");
-            assert!(!resp.artifacts.is_empty());
-        }
-
-        /// Real-API smoke for Lyria. Ignored by default.
-        #[test]
-        #[ignore = "needs GEMINI_API_KEY"]
-        fn gemini_lyria_smoke() {
-            use super::super::UreqGeminiAdapter;
-            use crate::contentgen::adapter::{GeminiAdapter, GeminiMusicRequest};
-            // Test-only: the live-API smoke reads an external credential
-            // (GEMINI_API_KEY), not cap config; gated `#[ignore]`.
-            #[allow(clippy::disallowed_methods)]
-            let key = env::var("GEMINI_API_KEY").expect("GEMINI_API_KEY set for smoke");
-            let adapter = UreqGeminiAdapter::new(key, Duration::from_mins(2));
-            let resp = adapter
-                .lyria_generate(GeminiMusicRequest {
-                    model: "lyria-3".to_string(),
-                    prompt: "calm ambient pad".to_string(),
-                    sample_count: 1,
-                })
-                .expect("live lyria request succeeds");
-            assert!(!resp.artifacts.is_empty());
-        }
-
-        fn request_nonce() -> u64 {
-            SystemTime::now().duration_since(UNIX_EPOCH).map_or(0, |d| {
-                #[allow(clippy::cast_possible_truncation)]
-                let n = d.as_nanos() as u64;
-                n
-            })
-        }
-
-        /// With every knob set, the request body carries
-        /// `imageConfig.imageSize`, `thinkingConfig.thinkingLevel` /
-        /// `includeThoughts`, and `tools[0].google_search` (issue 1167).
-        #[test]
-        fn nanobanana_body_carries_set_params() {
-            use crate::contentgen::adapter::GeminiImageRequest;
-            let body = super::super::build_nanobanana_body(&GeminiImageRequest {
-                model: "gemini-3.1-flash-image-preview".to_string(),
-                prompt: "a cat".to_string(),
-                aspect_ratio: "16:9".to_string(),
-                image_size: Some("2K".to_string()),
-                thinking_level: Some("high".to_string()),
-                include_thoughts: Some(true),
-                use_grounding: true,
-                reference_images: Vec::new(),
-            });
-            let gcfg = &body["generationConfig"];
-            assert_eq!(gcfg["imageConfig"]["aspectRatio"], "16:9");
-            assert_eq!(gcfg["imageConfig"]["imageSize"], "2K");
-            assert_eq!(gcfg["thinkingConfig"]["thinkingLevel"], "high");
-            assert_eq!(gcfg["thinkingConfig"]["includeThoughts"], true);
-            assert_eq!(body["tools"][0]["google_search"], serde_json::json!({}));
-        }
-
-        /// With the optional knobs unset, the body has no `imageSize`,
-        /// no `thinkingConfig`, and no `tools` key — only the always-on
-        /// `aspectRatio` survives under `imageConfig`.
-        #[test]
-        fn nanobanana_body_omits_unset_params() {
-            use crate::contentgen::adapter::GeminiImageRequest;
-            let body = super::super::build_nanobanana_body(&GeminiImageRequest {
-                model: "gemini-3.1-flash-image-preview".to_string(),
-                prompt: "a cat".to_string(),
-                aspect_ratio: "1:1".to_string(),
-                reference_images: Vec::new(),
-                ..Default::default()
-            });
-            let gcfg = &body["generationConfig"];
-            assert_eq!(gcfg["imageConfig"]["aspectRatio"], "1:1");
-            assert!(gcfg["imageConfig"].get("imageSize").is_none());
-            assert!(gcfg.get("thinkingConfig").is_none());
-            assert!(body.get("tools").is_none());
-        }
-
-        /// Serializes the two seam tests that pin `AETHER_GEN_DIR` so
-        /// their process-env mutation can't race nextest's other threads.
-        static GEN_DIR_ENV_LOCK: Mutex<()> = Mutex::new(());
-
-        /// Build a single-artifact `GeminiResponse` whose parse carried a
-        /// `thought_signature`, the shape the cap's reply assembly sees.
-        fn nb_response_with_signature(sig: &str) -> GeminiResponse {
-            GeminiResponse {
-                artifacts: vec![GeminiArtifact {
-                    bytes: STUB_PNG.to_vec(),
-                    ext: "png".to_string(),
-                }],
-                model_used: "gemini-3-pro-image-preview".to_string(),
-                usage: AdapterUsage::default(),
-                thought_signature: Some(sig.to_string()),
-                grounding: None,
-            }
-        }
-
-        /// Stage `nanobanana_reply` under a scratch `AETHER_GEN_DIR` so
-        /// the seam tests never touch the user's real save dir. Holds the
-        /// env lock across the set/reply/clear window.
-        fn reply_under_scratch_gen_dir(
-            include_sig: bool,
-            resp: GeminiResponse,
-        ) -> NanobananaGenerateResult {
-            let _guard = GEN_DIR_ENV_LOCK
-                .lock()
-                .unwrap_or_else(PoisonError::into_inner);
-            let scratch = env::temp_dir().join(format!(
-                "aether-gemini-sig-{}-{}",
-                process::id(),
-                request_nonce()
-            ));
-            fs::create_dir_all(&scratch).expect("scratch dir creates");
-            // SAFETY: serialized by GEN_DIR_ENV_LOCK against the other
-            // seam test; no other test reads this var.
-            unsafe {
-                env::set_var("AETHER_GEN_DIR", &scratch);
-            }
-            let reply = super::nanobanana_reply(1, include_sig, Ok(resp));
-            // SAFETY: same lock-guarded window as the set above.
-            unsafe {
-                env::remove_var("AETHER_GEN_DIR");
-            }
-            let _ = fs::remove_dir_all(&scratch);
-            reply
-        }
-
-        /// Default-off: a response carrying a `thought_signature` is
-        /// cleared from the reply when the flag is unset/false — the
-        /// fix for the multi-MB signature dominating the result.
-        #[test]
-        fn thought_signature_cleared_when_flag_off() {
-            let reply = reply_under_scratch_gen_dir(false, nb_response_with_signature("sig-abc"));
-            match reply {
-                NanobananaGenerateResult::Ok {
-                    thought_signature, ..
-                } => {
-                    assert_eq!(
-                        thought_signature, None,
-                        "flag off clears the signature from the reply"
-                    );
-                }
-                other @ NanobananaGenerateResult::Err { .. } => {
-                    panic!("expected Ok, got {other:?}")
-                }
-            }
-        }
-
-        /// Opt-in: the multi-turn continuation path is unaffected — with
-        /// the flag true the signature is retained exactly as parsed.
-        #[test]
-        fn thought_signature_retained_when_flag_on() {
-            let reply = reply_under_scratch_gen_dir(true, nb_response_with_signature("sig-abc"));
-            match reply {
-                NanobananaGenerateResult::Ok {
-                    thought_signature, ..
-                } => {
-                    assert_eq!(
-                        thought_signature.as_deref(),
-                        Some("sig-abc"),
-                        "flag on retains the signature for a multi-turn continuation"
-                    );
-                }
-                other @ NanobananaGenerateResult::Err { .. } => {
-                    panic!("expected Ok, got {other:?}")
-                }
-            }
-        }
-
-        /// The flag is cross-model, not an NB2-only knob: Pro accepts
-        /// `include_thought_signature: Some(true)` and dispatches rather
-        /// than rejecting with `MissingRequiredField`. Mirror of
-        /// `nb2_only_knob_rejected_on_older_model`, asserting acceptance.
-        #[test]
-        fn thought_signature_flag_accepted_on_pro() {
-            // Acceptance dispatches off-thread, so the reply lands
-            // asynchronously — peeking at the reply channel here would race
-            // that dispatch (iamacoffeepot/aether#1296). The deterministic
-            // proof of acceptance is the in-flight count, which `submit`
-            // bumps synchronously on this thread; a synchronous validation
-            // error `return`s before dispatch, leaving it at 0. So we don't
-            // need the reply channel at all.
-            let (mailer, _rx) = test_mailer_and_rx();
-            let cap_mailbox = MailboxId(0);
-            let mut cap = GeminiCapability::from_parts(Arc::new(StubGeminiAdapter), 4);
-            let transport = Arc::new(NativeBinding::new_for_test(
-                Arc::clone(&mailer),
-                cap_mailbox,
-            ));
-            let mut ctx = NativeCtx::new_dispatching(
-                &transport,
-                session_sender(),
-                aether_data::MailId::NONE,
-                aether_data::MailId::NONE,
-            );
-            let mut req = nb_request("gemini-3-pro-image-preview", AspectRatio::ASPECT_RATIO_1_1);
-            req.image_size = Some(ImageSize::K1);
-            req.include_thought_signature = Some(true);
-            cap.on_nanobanana_generate(&mut ctx, req);
-            assert_eq!(
-                cap.test_in_flight(),
-                1,
-                "Pro must accept the cross-model signature flag and dispatch \
-                 it rather than rejecting synchronously"
-            );
-        }
-    }
-}
